extract terminal view specific adaptors into a separate file
